finish this app

+++ lib/convert/data/models/convert_model.dart
@@ -1,7 +1,7 @@
-import 'package:qimahalan/core/databases/api/end_points.dart';
-import 'package:qimahalan/convert/data/models/sub_models/info_model.dart';
-import 'package:qimahalan/convert/data/models/sub_models/query_model.dart';
-import 'package:qimahalan/convert/domain/entities/convert_entities.dart';
+import 'package:Dinaro/core/databases/api/end_points.dart';
+import 'package:Dinaro/convert/data/models/sub_models/info_model.dart';
+import 'package:Dinaro/convert/data/models/sub_models/query_model.dart';
+import 'package:Dinaro/convert/domain/entities/convert_entities.dart';
 
 class ConvertModel extends ConvertEntities {
   const ConvertModel({
@@ -13,6 +13,6 @@ class ConvertModel extends ConvertEntities {
   factory ConvertModel.fromJson(Map<String, dynamic> json) => ConvertModel(
     info: InfoModel.fromJson(json[ApiKey.info]),
     query: QueryModel.fromJson(json[ApiKey.query]),
-    result: json[ApiKey.result],
+    result: json[ApiKey.result].toDouble(),
   );
 }
