+++ lib/convert/data/models/sub_models/info_model.dart
@@ -1,9 +1,9 @@
-import 'package:qimahalan/core/databases/api/end_points.dart';
-import 'package:qimahalan/convert/domain/entities/sub_entites/info_entites.dart';
+import 'package:Dinaro/core/databases/api/end_points.dart';
+import 'package:Dinaro/convert/domain/entities/sub_entites/info_entites.dart';
 
 class InfoModel extends InfoEntite {
   const InfoModel({required super.quote});
 
   factory InfoModel.fromJson(Map<String, dynamic> json) =>
-      InfoModel(quote: json[ApiKey.quote]);
+      InfoModel(quote: json[ApiKey.quote].toDouble());
 }
